+++ impacted-test-engine/src/main/java/com/teamscale/test_impacted/engine/ImpactedTestEngine.java
@@ -1,5 +1,6 @@
 package com.teamscale.test_impacted.engine;
 
+import com.teamscale.test_impacted.commons.LoggerUtils;
 import com.teamscale.test_impacted.engine.options.TestEngineOptionUtils;
 import com.teamscale.test_impacted.engine.options.TestEngineOptions;
 import org.junit.platform.engine.EngineDiscoveryRequest;
@@ -8,11 +9,16 @@ import org.junit.platform.engine.TestDescriptor;
 import org.junit.platform.engine.TestEngine;
 import org.junit.platform.engine.UniqueId;
 
+import java.util.Optional;
+import java.util.logging.Logger;
+
 /** Test engine for executing impacted tests. */
 public class ImpactedTestEngine implements TestEngine {
 
 	/** The id of the {@link ImpactedTestEngine}. */
 	public static final String ENGINE_ID = "teamscale-test-impacted";
+
+	public static final Logger LOGGER = LoggerUtils.getLogger(ImpactedTestEngine.class);
 
 	private InternalImpactedTestEngine internalImpactedTestEngine = null;
 
@@ -29,19 +35,28 @@ public class ImpactedTestEngine implements TestEngine {
 
 		// Re-initialize the configuration for this discovery (and optional following execution).
 		internalImpactedTestEngine =
-				new InternalImpactedTestEngine(configuration.testEngineRegistry, configuration.testExecutor,
-						new TestDataWriter(configuration.reportDirectory), engineOptions.getPartition());
+				new InternalImpactedTestEngine(configuration, engineOptions.getPartition());
 
 		return internalImpactedTestEngine.discover(discoveryRequest, uniqueId);
 	}
 
 	@Override
 	public void execute(ExecutionRequest request) {
-		// According to the TestEngine interface the request must correspond to the last execution request. Therefore we
+		// According to the TestEngine interface the request must correspond to the last execution request. Therefore, we
 		// may re-use the configuration initialized during discovery.
 		if (internalImpactedTestEngine == null) {
 			throw new AssertionError("Can't execute request without discovering it first.");
 		}
 		internalImpactedTestEngine.execute(request);
+	}
+
+	@Override
+	public Optional<String> getGroupId() {
+		return Optional.of("com.teamscale");
+	}
+
+	@Override
+	public Optional<String> getArtifactId() {
+		return Optional.of("impacted-test-engine");
 	}
 }