+++ impacted-test-engine/src/main/java/com/teamscale/test_impacted/engine/InternalImpactedTestEngine.java
@@ -1,19 +1,18 @@
 package com.teamscale.test_impacted.engine;
 
-import com.teamscale.client.TestDetails;
 import com.teamscale.report.testwise.model.TestExecution;
+import com.teamscale.test_impacted.commons.LoggerUtils;
 import com.teamscale.test_impacted.engine.executor.AvailableTests;
-import com.teamscale.test_impacted.engine.executor.ITestExecutor;
-import com.teamscale.test_impacted.engine.executor.TestExecutorRequest;
+import com.teamscale.test_impacted.engine.executor.ITestSorter;
+import com.teamscale.test_impacted.engine.executor.TeamscaleAgentNotifier;
+import com.teamscale.test_impacted.engine.executor.TestwiseCoverageCollectingExecutionListener;
+import com.teamscale.test_impacted.test_descriptor.ITestDescriptorResolver;
+import com.teamscale.test_impacted.test_descriptor.TestDescriptorResolverRegistry;
 import com.teamscale.test_impacted.test_descriptor.TestDescriptorUtils;
-import org.junit.platform.commons.logging.Logger;
-import org.junit.platform.commons.logging.LoggerFactory;
 import org.junit.platform.engine.EngineDiscoveryRequest;
-import org.junit.platform.engine.EngineExecutionListener;
 import org.junit.platform.engine.ExecutionRequest;
 import org.junit.platform.engine.TestDescriptor;
 import org.junit.platform.engine.TestEngine;
-import org.junit.platform.engine.TestExecutionResult;
 import org.junit.platform.engine.UniqueId;
 import org.junit.platform.engine.support.descriptor.EngineDescriptor;
 
@@ -21,6 +20,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.ServiceLoader;
+import java.util.logging.Logger;
 
 import static com.teamscale.test_impacted.engine.ImpactedTestEngine.ENGINE_ID;
 
@@ -29,22 +29,23 @@ import static com.teamscale.test_impacted.engine.ImpactedTestEngine.ENGINE_ID;
  */
 class InternalImpactedTestEngine {
 
-	private static final Logger LOGGER = LoggerFactory.getLogger(InternalImpactedTestEngine.class);
+	private static final Logger LOGGER = LoggerUtils.getLogger(InternalImpactedTestEngine.class);
 
 	private final TestEngineRegistry testEngineRegistry;
 
-	private final ITestExecutor testExecutor;
+	private final ITestSorter testSorter;
+
+	private final TeamscaleAgentNotifier teamscaleAgentNotifier;
 
 	private final TestDataWriter testDataWriter;
 
 	private final String partition;
 
-	InternalImpactedTestEngine(TestEngineRegistry testEngineRegistry,
-							   ITestExecutor testExecutor,
-							   TestDataWriter testDataWriter, String partition) {
-		this.testEngineRegistry = testEngineRegistry;
-		this.testExecutor = testExecutor;
-		this.testDataWriter = testDataWriter;
+	InternalImpactedTestEngine(ImpactedTestEngineConfiguration configuration, String partition) {
+		this.testEngineRegistry = configuration.testEngineRegistry;
+		this.testSorter = configuration.testSorter;
+		this.testDataWriter = configuration.testDataWriter;
+		this.teamscaleAgentNotifier = configuration.teamscaleAgentNotifier;
 		this.partition = partition;
 	}
 
@@ -55,17 +56,17 @@ class InternalImpactedTestEngine {
 	TestDescriptor discover(EngineDiscoveryRequest discoveryRequest, UniqueId uniqueId) {
 		EngineDescriptor engineDescriptor = new EngineDescriptor(uniqueId, "Teamscale Impacted Tests");
 
-		LOGGER.debug(() -> "Starting test discovery for engine " + ENGINE_ID);
+		LOGGER.fine(() -> "Starting test discovery for engine " + ENGINE_ID);
 
 		for (TestEngine delegateTestEngine : testEngineRegistry) {
-			LOGGER.debug(() -> "Starting test discovery for delegate engine: " + delegateTestEngine.getId());
+			LOGGER.fine(() -> "Starting test discovery for delegate engine: " + delegateTestEngine.getId());
 			TestDescriptor delegateEngineDescriptor = delegateTestEngine.discover(discoveryRequest,
 					UniqueId.forEngine(delegateTestEngine.getId()));
 
 			engineDescriptor.addChild(delegateEngineDescriptor);
 		}
 
-		LOGGER.debug(() -> "Discovered test descriptor for engine " + ENGINE_ID + ":\n" + TestDescriptorUtils
+		LOGGER.fine(() -> "Discovered test descriptor for engine " + ENGINE_ID + ":\n" + TestDescriptorUtils
 				.getTestDescriptorAsString(engineDescriptor));
 
 		return engineDescriptor;
@@ -76,42 +77,47 @@ class InternalImpactedTestEngine {
 	 * {@link #discover(EngineDiscoveryRequest, UniqueId)} with the corresponding {@link TestEngine}.
 	 */
 	void execute(ExecutionRequest request) {
-		EngineExecutionListener engineExecutionListener = request.getEngineExecutionListener();
-		EngineDescriptor engineDescriptor = (EngineDescriptor) request.getRootTestDescriptor();
+		TestDescriptor rootTestDescriptor = request.getRootTestDescriptor();
+		AvailableTests availableTests = TestDescriptorUtils.getAvailableTests(rootTestDescriptor, partition);
 
-		LOGGER.debug(() -> "Starting execution of request for engine " + ENGINE_ID + ":\n" + TestDescriptorUtils
-				.getTestDescriptorAsString(engineDescriptor));
+		LOGGER.fine(() -> "Starting selection and sorting " + ENGINE_ID + ":\n" + TestDescriptorUtils
+				.getTestDescriptorAsString(rootTestDescriptor));
 
-		engineExecutionListener.executionStarted(engineDescriptor);
-		runTestExecutor(request);
-		engineExecutionListener.executionFinished(engineDescriptor, TestExecutionResult.successful());
+		testSorter.selectAndSort(rootTestDescriptor);
+
+		LOGGER.fine(() -> "Starting execution of request for engine " + ENGINE_ID + ":\n" + TestDescriptorUtils
+				.getTestDescriptorAsString(rootTestDescriptor));
+
+		List<TestExecution> testExecutions = executeTests(request, rootTestDescriptor);
+
+		testDataWriter.dumpTestExecutions(testExecutions);
+		testDataWriter.dumpTestDetails(availableTests.getTestList());
+		teamscaleAgentNotifier.testRunEnded();
 	}
 
-	private void runTestExecutor(ExecutionRequest request) {
-		List<TestDetails> availableTests = new ArrayList<>();
+	private List<TestExecution> executeTests(ExecutionRequest request, TestDescriptor rootTestDescriptor) {
 		List<TestExecution> testExecutions = new ArrayList<>();
-
-		for (TestDescriptor engineTestDescriptor : request.getRootTestDescriptor().getChildren()) {
+		for (TestDescriptor engineTestDescriptor : rootTestDescriptor.getChildren()) {
 			Optional<String> engineId = engineTestDescriptor.getUniqueId().getEngineId();
-
 			if (!engineId.isPresent()) {
-				LOGGER.error(
-						() -> "Engine id for test descriptor " + engineTestDescriptor + " not present. Skipping execution of the engine.");
+				LOGGER.severe(
+						() -> "Engine ID for test descriptor " + engineTestDescriptor + " not present. Skipping execution of the engine.");
 				continue;
 			}
 
 			TestEngine testEngine = testEngineRegistry.getTestEngine(engineId.get());
-			AvailableTests availableTestsForEngine = TestDescriptorUtils
-					.getAvailableTests(testEngine, engineTestDescriptor, partition);
-			TestExecutorRequest testExecutorRequest = new TestExecutorRequest(testEngine, engineTestDescriptor,
-					request.getEngineExecutionListener(), request.getConfigurationParameters());
-			List<TestExecution> testExecutionsOfEngine = testExecutor.execute(testExecutorRequest);
+			ITestDescriptorResolver testDescriptorResolver = TestDescriptorResolverRegistry
+					.getTestDescriptorResolver(testEngine.getId());
+			TestwiseCoverageCollectingExecutionListener executionListener =
+					new TestwiseCoverageCollectingExecutionListener(teamscaleAgentNotifier,
+							testDescriptorResolver,
+							request.getEngineExecutionListener());
 
-			testExecutions.addAll(testExecutionsOfEngine);
-			availableTests.addAll(availableTestsForEngine.getTestList());
+			testEngine.execute(new ExecutionRequest(engineTestDescriptor, executionListener,
+					request.getConfigurationParameters()));
+
+			testExecutions.addAll(executionListener.getTestExecutions());
 		}
-
-		testDataWriter.dumpTestDetails(availableTests);
-		testDataWriter.dumpTestExecutions(testExecutions);
+		return testExecutions;
 	}
 }