+++ impacted-test-engine/src/main/java/com/teamscale/test_impacted/engine/options/TestEngineOptions.java
@@ -4,11 +4,13 @@ import com.teamscale.client.CommitDescriptor;
 import com.teamscale.client.TeamscaleClient;
 import com.teamscale.test_impacted.engine.ImpactedTestEngine;
 import com.teamscale.test_impacted.engine.ImpactedTestEngineConfiguration;
+import com.teamscale.test_impacted.engine.TestDataWriter;
 import com.teamscale.test_impacted.engine.TestEngineRegistry;
-import com.teamscale.test_impacted.engine.executor.ITestExecutor;
-import com.teamscale.test_impacted.engine.executor.ImpactedTestsExecutor;
+import com.teamscale.test_impacted.engine.executor.ITestSorter;
+import com.teamscale.test_impacted.engine.executor.ImpactedTestsSorter;
 import com.teamscale.test_impacted.engine.executor.ImpactedTestsProvider;
-import com.teamscale.test_impacted.engine.executor.TestwiseCoverageCollectingTestExecutor;
+import com.teamscale.test_impacted.engine.executor.NOPTestSorter;
+import com.teamscale.test_impacted.engine.executor.TeamscaleAgentNotifier;
 import com.teamscale.tia.client.ITestwiseCoverageAgentApi;
 import okhttp3.HttpUrl;
 import org.junit.platform.engine.TestEngine;
@@ -56,15 +58,13 @@ public class TestEngineOptions {
 	private List<ITestwiseCoverageAgentApi> testwiseCoverageAgentApis = Collections.emptyList();
 
 	/** The test engine ids of all {@link TestEngine}s to use. If empty all available {@link TestEngine}s are used. */
-	private Set<String> testEngineIds = Collections.emptySet();
+	private Set<String> includedTestEngineIds = Collections.emptySet();
+
+	/** The test engine ids of all {@link TestEngine}s to exclude. */
+	private Set<String> excludedTestEngineIds = Collections.emptySet();
 
 	/** The directory used to store test-wise coverage reports. Must be a writeable directory. */
 	private File reportDirectory;
-
-	/** @see #runImpacted */
-	private boolean isRunImpacted() {
-		return runImpacted;
-	}
 
 	/** @see #runAllTests */
 	private boolean isRunAllTests() {
@@ -87,23 +87,34 @@ public class TestEngineOptions {
 	}
 
 	public ImpactedTestEngineConfiguration createTestEngineConfiguration() {
-		ITestExecutor testExecutor = createTestExecutor();
-		TestEngineRegistry testEngineRegistry = new TestEngineRegistry(testEngineIds);
+		ITestSorter testSorter = createTestSorter();
+		TeamscaleAgentNotifier teamscaleAgentNotifier = createTeamscaleAgentNotifier();
+		TestEngineRegistry testEngineRegistry = new TestEngineRegistry(includedTestEngineIds, excludedTestEngineIds);
+		TestDataWriter testDataWriter = new TestDataWriter(reportDirectory);
 
-		return new ImpactedTestEngineConfiguration(reportDirectory, testEngineRegistry, testExecutor);
+		return new ImpactedTestEngineConfiguration(testDataWriter, testEngineRegistry, testSorter, teamscaleAgentNotifier);
 	}
 
-	private ITestExecutor createTestExecutor() {
-		if (!isRunImpacted()) {
-			return new TestwiseCoverageCollectingTestExecutor(testwiseCoverageAgentApis, false);
+	private ITestSorter createTestSorter() {
+		if (!runImpacted) {
+			return new NOPTestSorter();
 		}
 
+		ImpactedTestsProvider testsProvider = createImpactedTestsProvider();
+		return new ImpactedTestsSorter(testsProvider);
+	}
+
+	private ImpactedTestsProvider createImpactedTestsProvider() {
 		TeamscaleClient client = new TeamscaleClient(serverOptions.getUrl(), serverOptions.getUserName(),
 				serverOptions.getUserAccessToken(), serverOptions.getProject(),
 				new File(reportDirectory, "server-request.txt"));
-		ImpactedTestsProvider testsProvider = new ImpactedTestsProvider(client, baseline, endCommit, partition,
+		return new ImpactedTestsProvider(client, baseline, endCommit, partition,
 				isRunAllTests(), isIncludeAddedTests(), isIncludeFailedAndSkipped());
-		return new ImpactedTestsExecutor(testwiseCoverageAgentApis, testsProvider);
+	}
+
+	private TeamscaleAgentNotifier createTeamscaleAgentNotifier() {
+		return new TeamscaleAgentNotifier(testwiseCoverageAgentApis,
+				runImpacted && !runAllTests);
 	}
 
 	/** Returns the builder for {@link TestEngineOptions}. */
@@ -177,9 +188,15 @@ public class TestEngineOptions {
 			return this;
 		}
 
-		/** @see #testEngineIds */
-		public Builder testEngineIds(List<String> testEngineIds) {
-			testEngineOptions.testEngineIds = new HashSet<>(testEngineIds);
+		/** @see #includedTestEngineIds */
+		public Builder includedTestEngineIds(List<String> testEngineIds) {
+			testEngineOptions.includedTestEngineIds = new HashSet<>(testEngineIds);
+			return this;
+		}
+
+		/** @see #excludedTestEngineIds */
+		public Builder excludedTestEngineIds(List<String> testEngineIds) {
+			testEngineOptions.excludedTestEngineIds = new HashSet<>(testEngineIds);
 			return this;
 		}