+++ system-tests/artifactory-git-properties-detection/src/test/java/com/teamscale/client/ArtifactoryGitPropertiesDetectionTest.java
@@ -5,21 +5,17 @@ import org.conqat.lib.commons.collections.PairList;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
 import systemundertest.SystemUnderTest;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertAll;
 
 public class ArtifactoryGitPropertiesDetectionTest {
 
 	/**
 	 * This port must match what is configured for the -javaagent line in the build.gradle file.
 	 */
-	private static final int FAKE_ARTIFACTORY_PORT = 65432;
-	private static final int AGENT_PORT = 65433;
+	private static final int FAKE_ARTIFACTORY_PORT = 63100;
+	private static final int AGENT_PORT = 63101;
 	private ArtifactoryMockServer artifactoryMockServer = null;
 
 	@BeforeEach