+++ system-tests/default-excludes-test/src/test/java/com/teamscale/tia/client/DefaultExcludesSystemTest.java
@@ -14,8 +14,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 public class DefaultExcludesSystemTest {
 
 	/** These ports must match what is configured for the -javaagent line in this project's build.gradle. */
-	private static final int FAKE_TEAMSCALE_PORT = 65437;
-	private static final int AGENT_PORT = 65436;
+	private static final int FAKE_TEAMSCALE_PORT = 63200;
+	private static final int AGENT_PORT = 63201;
 
 	@Test
 	public void systemTest() throws Exception {
@@ -28,7 +28,7 @@ public class DefaultExcludesSystemTest {
 		SystemTestUtils.dumpCoverage(AGENT_PORT);
 
 		assertThat(teamscaleMockServer.uploadedReports).hasSize(1);
-		String report = teamscaleMockServer.uploadedReports.get(0);
+		String report = teamscaleMockServer.uploadedReports.get(0).getReportString();
 		assertThat(report).doesNotContain("shadow");
 		assertThat(report).doesNotContain("junit");
 		assertThat(report).doesNotContain("eclipse");